+++ Cargo.toml
@@ -7,5 +7,6 @@ members = [
     "gamercade_core",
     "gamercade_editor",
     "gamercade_rs",
+    "gamercade_sound_engine",
     "gamercade_tools",
 ]
