+++ gamercade_audio/src/instruments/sampler/mod.rs
@@ -1,9 +1,5 @@
 mod sample_definition;
-mod sample_oscillator;
-mod sampler_instance;
 
 pub use sample_definition::*;
-pub use sample_oscillator::*;
-pub use sampler_instance::*;
 
 pub type SampleBitDepth = i16;
