+++ gamercade_audio/src/instruments/wavetable/mod.rs
@@ -1,11 +1,9 @@
 mod wavetable_definition;
 mod wavetable_generator;
-mod wavetable_instance;
 mod wavetable_waveform;
 
 pub use wavetable_definition::*;
 pub use wavetable_generator::*;
-pub use wavetable_instance::*;
 pub use wavetable_waveform::*;
 
 pub type WavetableBitDepth = i16;
