+++ gamercade_audio/src/lib.rs
@@ -1,41 +1,13 @@
 mod consts;
-mod envelope;
+mod envelope_definition;
 mod instruments;
 mod notes;
 mod sound_rom;
 mod tracker;
 
 pub use consts::*;
-pub use envelope::*;
+pub use envelope_definition::*;
 pub use instruments::*;
 pub use notes::*;
 pub use sound_rom::*;
 pub use tracker::*;
-
-#[cfg(feature = "playback")]
-mod playback;
-#[cfg(feature = "playback")]
-mod sound_engine;
-#[cfg(feature = "playback")]
-mod sound_rom_instance;
-
-#[cfg(feature = "playback")]
-pub use playback::*;
-#[cfg(feature = "playback")]
-pub use sound_engine::*;
-#[cfg(feature = "playback")]
-pub use sound_rom_instance::*;
-
-pub const MAX_ROLLBACK_SOUNDS: usize = 8;
-
-#[cfg(feature = "playback")]
-fn initialize_globals() {
-    init_fm_lut();
-    initialize_notes();
-    unsafe {
-        NO_SOUND_DEFINITION.write(std::sync::Arc::new(WavetableDefinition {
-            data: Box::new([0, 0]),
-            envelope: EnvelopeDefinition::default(),
-        }));
-    }
-}
