+++ gamercade_editor/Cargo.toml
@@ -8,7 +8,8 @@ edition = "2021"
 [dependencies]
 # Gamercade Dependencies
 gamercade_core = { path = "../gamercade_core" }
-gamercade_audio = { path = "../gamercade_audio", features = ["playback"] }
+gamercade_audio = { path = "../gamercade_audio" }
+gamercade_sound_engine = { path = "../gamercade_sound_engine" }
 
 # Serialization
 serde = { version = "1.0.144", features = ["derive"] }
