+++ gamercade_sound_engine/src/envelope/linear_ramp.rs
