+++ gamercade_sound_engine/src/envelope/mod.rs
@@ -1,0 +1,9 @@
+mod envelope_instance;
+mod envelope_phase;
+mod exponential_ramp;
+mod linear_ramp;
+
+pub use envelope_instance::*;
+pub(crate) use envelope_phase::*;
+pub use exponential_ramp::*;
+pub use linear_ramp::*;
