+++ gamercade_sound_engine/src/instruments/sampler/mod.rs
@@ -1,0 +1,5 @@
+mod sample_oscillator;
+mod sampler_instance;
+
+pub use sample_oscillator::*;
+pub use sampler_instance::*;
