+++ gamercade_sound_engine/src/instruments/wavetable/mod.rs
@@ -1,0 +1,5 @@
+mod wavetable_instance;
+mod wavetable_oscillator;
+
+pub use wavetable_instance::*;
+pub use wavetable_oscillator::*;
