+++ gamercade_sound_engine/src/playback/mod.rs
@@ -2,8 +2,12 @@ mod chain_playback;
 mod phrase_playback;
 mod sfx_playback;
 mod song_playback;
+mod tracker_flow;
+mod tracker_oscillator;
 
 pub use chain_playback::*;
 pub use phrase_playback::*;
 pub use sfx_playback::*;
 pub use song_playback::*;
+pub use tracker_flow::*;
+pub(crate) use tracker_oscillator::*;
